requirements.txt
cython
numpy
pyyaml
scipy
tensorboard
easydict
tqdm
